Melhorias e Organização do CSS
- Consolidado e organizado regras de CSS para vários componentes.
- Introduzidas variáveis CSS para cores primárias e de texto.
- Removidos seletores de filho direto desnecessários para maior flexibilidade.
- Agrupamento de propriedades relacionadas para melhor legibilidade e manutenção.
- Aplicação de estilos consistentes em todo o projeto, facilitando futuras atualizações de tema.

+++ admin/gerenciar_newsletter.php
@@ -1,45 +1,46 @@
 <?php
 include('auth_check.php');
 include('../config.php');
-session_start();?>
+session_start();
 
-// Verificar autenticação e permissões...
+// Verificar autenticação e permissões
+if (!isset($_SESSION['user_id']) || $_SESSION['is_admin'] != 1) {
+    header('Location: ../login.php');
+    exit();
+}
+?>
 
 <!DOCTYPE html>
 <html lang="pt-br">
-
 <head>
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <link rel="stylesheet" type="text/css" href="../css/admin.css" media="screen" />
-  <title>Newsletter</title>
+  <title>Gerenciar Assinaturas da Newsletter</title>
 </head>
-
 <body>
-  <?php include('../inc/adminHeader.php');?>
+  <?php include('../inc/adminHeader.php'); ?>
   <main class="admin__container">
+    <h2 class="title">Gerenciar Assinaturas da Newsletter</h2>
     <?php 
-    echo "<h2 class='title'>Gerenciar Assinaturas da Newsletter</h2>";
-
     $sql = "SELECT id_assinatura, email, preferencias, status FROM AssinaturasNewsletter";
     $stmt = $pdo->query($sql);
 
     echo "<table class='table'><tr><th class='name'>ID</th><th class='name'>Email</th><th class='name'>Preferências</th><th class='name'>Status</th><th class='name'>Ações</th></tr>";
     while ($row = $stmt->fetch(PDO::FETCH_ASSOC)) {
         echo "<tr class='table__wrapper'>
-                <td class='row id'>{$row['id_assinatura']}</td>
-                <td class='row'>{$row['email']}</td>
-                <td class='row'>{$row['preferencias']}</td>
-                <td class='row'>{$row['status']}</td>
+                <td class='row id'>" . htmlspecialchars($row['id_assinatura']) . "</td>
+                <td class='row'>" . htmlspecialchars($row['email']) . "</td>
+                <td class='row'>" . htmlspecialchars($row['preferencias']) . "</td>
+                <td class='row'>" . htmlspecialchars($row['status']) . "</td>
                 <td class='row buttons'>
-                    <a class='table__buttons remove' href='cancelar_assinatura.php?id={$row['id_assinatura']}' onclick='return confirm(\"Tem certeza?\");'>Cancelar Assinatura</a>
+                    <a class='table__buttons remove' href='cancelar_assinatura.php?id=" . htmlspecialchars($row['id_assinatura']) . "' onclick='return confirm(\"Tem certeza?\");'>Cancelar Assinatura</a>
                 </td>
             </tr>";
     }
     echo "</table>";
-?>
+    ?>
   </main>
-  <?php include('../inc/adminFooter.php');?>
+  <?php include('../inc/adminFooter.php'); ?>
 </body>
-
 </html>
